ML_service/requirements.txt
fastapi
uvicorn
openai
scikit-learn
pandas
PyMuPDF
pydantic
python-dotenv
requests
python-multipart
numpy
scipy
